+++ core/domain/src/main/java/team/retum/usecase/usecase/interests/SetInterestsToggleUseCase.kt
@@ -6,7 +6,7 @@ import javax.inject.Inject
 class SetInterestsToggleUseCase @Inject constructor(
     private val interestsRepository: InterestsRepository,
 ) {
-    suspend operator fun invoke(codes: List<Int>) = runCatching {
+    suspend operator fun invoke(codes: List<Long>) = runCatching {
         interestsRepository.setInterestsToggle(codes = codes)
     }
 }
